refactor resources, remove refresh tokens, extended token expiry time, changed token identity to primary key
