+++ server/run.py
@@ -1,4 +1,4 @@
-# to avoid circular import must save with imort views, models, resources at bottom
+# to avoid circular import, must save with import views, models, resources at bottom
 # to save without auto-formatting on vs code, `CTRL + K + SHIFT + S`
 
 from flask import Flask
@@ -20,15 +20,15 @@ db = SQLAlchemy(app)
 
 @app.before_first_request
 def create_tables():
-  db.create_all()
+    db.create_all()
 
 app.config['JWT_SECRET_KEY'] = os.environ.get('JWT_SECRET_KEY')
 jwt = JWTManager(app)
 
 # enable blacklisting in configuration
 app.config['JWT_BLACKLIST_ENABLED'] = True
-# specify what kind of token to check, access and refresh
-app.config['JWT_BLACKLIST_TOKEN_CHECKS'] = ['access', 'refresh']
+# specify what kind of token to check, access
+app.config['JWT_BLACKLIST_TOKEN_CHECKS'] = ['access']
 
 @jwt.token_in_blacklist_loader
 def check_if_token_in_blacklist(decrypted_token):
@@ -38,12 +38,10 @@ def check_if_token_in_blacklist(decrypted_token):
     return models.RevokedTokenModel.is_jti_blacklisted(jti)
 
 
-import views, models, resources
+import models, resource.secretResource, resource.allUsers, resource.userLogoutAccess, resource.userLogin, resource.userRegistration
 
-api.add_resource(resources.UserRegistration, '/registration')
-api.add_resource(resources.UserLogin, '/login')
-api.add_resource(resources.UserLogoutAccess, '/logout/access')
-api.add_resource(resources.UserLogoutRefresh, '/logout/refresh')
-api.add_resource(resources.TokenRefresh, '/token/refresh')
-api.add_resource(resources.AllUsers, '/users')
-api.add_resource(resources.SecretResource, '/secret')
+api.add_resource(resource.userRegistration.UserRegistration, '/registration')
+api.add_resource(resource.userLogin.UserLogin, '/login')
+api.add_resource(resource.userLogoutAccess.UserLogoutAccess, '/logout/access')
+api.add_resource(resource.allUsers.AllUsers, '/users')
+api.add_resource(resource.secretResource.SecretResource, '/secret')
