--- server/views.py
@@ -1,7 +1,0 @@
-from run import app
-from flask import jsonify
-
-
-@app.route('/')
-def index():
-    return jsonify({'message': 'Hello, World!'})
